// lib/models/panels.ts
export enum EPanels {
  TEXT = "TEXT",
  SHAPES = "SHAPES",
  IMAGES = "IMAGES",
  TEMPLATES = "TEMPLATES",
  BookCovers = "BookCovers",
  UPLOADS = "UPLOADS",
  BACKGROUND = "BACKGROUND",
  PAGES = "PAGES",
  PAGE_SETTINGS = "PAGE_SETTINGS",
  PROJECTS = "PROJECTS",
  ADMIN = "ADMIN",
  NONE = "NONE",
}